+++ packages/pigeon/lib/src/ast.dart
@@ -830,6 +830,8 @@ class Root extends Node {
     );
   }
 
+  // TODO(tarrinneal): Ensure classes are sorted in topological dependency order; see
+  // https://github.com/flutter/flutter/issues/128330.
   /// All the classes contained in the AST.
   List<Class> classes;