+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.1
+
+* Adds support to overlay billing related messages. See `InAppPurchaseAndroidPlatformAddition.showInAppMessages`.
+
 ## 0.5.0+2
 
 * Bumps the androidx group across 10 directories with 1 update.